openbsd: create mem.c module
