+++ setup.py
@@ -222,6 +222,7 @@ elif OPENBSD:
             'psutil/_psutil_bsd.c',
             'psutil/arch/openbsd/cpu.c',
             'psutil/arch/openbsd/disk.c',
+            'psutil/arch/openbsd/mem.c',
             'psutil/arch/openbsd/specific.c',
         ],
         define_macros=macros,
